ml-server/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
# Add ML dependencies as needed:
# torch>=2.0.0
# transformers>=4.30.0
# numpy>=1.24.0
# scikit-learn>=1.3.0
